Add files via upload
This is an application practice of Bosch Internet of things platform. We used Bosch Internet of things suite XDK and PPM system to complete a multiple fault detection and message prompt of conveyor belt in industrial environment. In order to better realize the fault monitoring, we have made a lot of efforts, including the selection of appropriate sensors and the analysis of some characteristics when and before the fault occurs. Although we have made a lot of preparations, the real conveyor belt demonstration model is still far more than we expected, its space is small, it is not convenient to install the optical sensor (wired) we prepared before, and when this actual extension of XDK's function is uploaded to the PPM platform, the opening of this data link also consumes a lot of energy. We share the original code of this simple practice with interested friends, hoping to have some help. At the same time, we also welcome you to correct it.
This time the original code is divided into four projects, which are the realization of the functions of the four XDK devices we do. In order to improve the accuracy of fault information collection, we improve the acquisition frequency of the sensor, and simplify the format and quantity of data upload according to different functions, and further improve the transmission rate.